+++ tests/wires/test_contour.py
@@ -8,7 +8,7 @@ import volmdlr.wires
 class TestContour(unittest.TestCase):
 
     def test_is_overlapping(self):
-        p1 = [volmdlr.Point2D(-0.2, -0.2), volmdlr.Point2D(0.2, -0.2), 
+        p1 = [volmdlr.Point2D(-0.2, -0.2), volmdlr.Point2D(0.2, -0.2),
               volmdlr.Point2D(0.2, 0.2), volmdlr.Point2D(-0.2, 0.2)]
 
         contour1 = volmdlr.wires.ClosedPolygon2D(p1)