+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * OpenShell3D: project_coincident_faces_of
 * GmshParser: to_vtk
 * BSplineCurve: derivatives
+* ArcEllipse2D: line_intersections, frame_mapping, linesegment_intersections
 
 
 ### Fixed
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Step: fix reading of rational BSpline curves and surfaces from step file.
 * BSplineCurve2D: tangent (use position/length)
 * Babylon: some scene settings for better rendering
+
 
 
 ### Removed
@@ -50,6 +52,7 @@ local_to_global_coordinates and global_to_local_coordinates are the new more exp
 
 * Contour2D: point_belongs
 * Basis2D, Basis3D, Frame2D, Frame3D: local_to_global_coordinates and global_to_local_coordinates
+* ArcEllipse2D: linesegment_intersections